+++ xarray/coding/cftimeindex.py
@@ -260,6 +260,7 @@ def format_attrs(index, separator=", "):
         "dtype": f"'{index.dtype}'",
         "length": f"{len(index)}",
         "calendar": f"'{index.calendar}'",
+        "freq": f"'{index.freq}'"
     }
     attrs_str = [f"{k}={v}" for k, v in attrs.items()]
     attrs_str = f"{separator}".join(attrs_str)
@@ -682,6 +683,12 @@ class CFTimeIndex(pd.Index):
         from .times import infer_calendar_name
 
         return infer_calendar_name(self)
+
+    @property
+    def freq(self):
+        """The frequency used by the dates in the index."""
+        from .frequencies import infer_freq
+        return infer_freq(self)
 
     def _round_via_method(self, freq, method):
         """Round dates using a specified method."""