+++ xarray/tests/test_cftimeindex.py
@@ -944,6 +944,12 @@ def test_cftimeindex_periods_repr(periods):
     repr_str = index.__repr__()
     assert f" length={periods}" in repr_str
 
+@requires_cftime
+@pytest.mark.parametrize("calendar", ["noleap", "360_day"])
+@pytest.mark.parametrize("freq", ["1D", "MS"])
+def test_cftimeindex_freq_in_repr(freq,calendar):
+    index = xr.cftime_range(start="2000", periods=3, freq=freq, calendar=calendar)
+    assert f', freq={freq}' in index.__repr__()
 
 @requires_cftime
 @pytest.mark.parametrize(